requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.6.1=pyhd8ed1ab_1
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310he2412df_2
asttokens=2.0.8=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.9.24=h5b45459_0
cffi=1.15.1=py310hcbf9ad4_0
colorama=0.4.5=pyhd8ed1ab_0
debugpy=1.6.3=py310h8a704f9_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.1.1=pyhd8ed1ab_0
flit-core=3.7.1=pyhd8ed1ab_0
gettext=0.21.1=h5728263_0
glib=2.74.0=h12be248_0
glib-tools=2.74.0=h12be248_0
gst-plugins-base=1.20.3=h001b923_2
gstreamer=1.20.3=h6b5321d_2
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310h5588dad_0
importlib_resources=5.10.0=pyhd8ed1ab_0
ipykernel=6.16.0=pyh025b116_0
ipython=8.5.0=pyh08f2357_1
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.2=pyhd8ed1ab_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h8ffe710_2
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter=1.0.0=py310h5588dad_7
jupyter_client=7.4.2=pyhd8ed1ab_0
jupyter_console=6.4.4=pyhd8ed1ab_0
jupyter_core=4.11.1=py310h5588dad_0
jupyter_server=1.21.0=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.3=pyhd8ed1ab_0
krb5=1.19.3=h1176d77_0
libclang=14.0.6=default_hf4acfee_0
libclang13=14.0.6=default_h8992b79_0
libffi=3.4.2=h8ffe710_5
libglib=2.74.0=h79619a9_0
libiconv=1.17=h8ffe710_0
libogg=1.3.4=h8ffe710_1
libpng=1.6.38=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.39.4=hcfcfb64_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libzlib=1.2.13=hcfcfb64_4
markupsafe=2.1.1=py310he2412df_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
nbclassic=0.4.5=pyhd8ed1ab_0
nbclient=0.7.0=pyhd8ed1ab_0
nbconvert=7.2.1=pyhd8ed1ab_0
nbconvert-core=7.2.1=pyhd8ed1ab_0
nbconvert-pandoc=7.2.1=pyhd8ed1ab_0
nbformat=5.7.0=pyhd8ed1ab_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
notebook=6.5.1=pyha770c72_0
notebook-shim=0.2.0=pyhd8ed1ab_0
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandoc=2.19.2=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.37=hdfff0fc_1
pickleshare=0.7.5=py_1003
pip=22.3=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
ply=3.11=py_1
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
prompt_toolkit=3.0.31=hd8ed1ab_0
psutil=5.9.3=py310h8d17308_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pypdf2=2.10.8=pyhd8ed1ab_0
pyqt=5.15.7=py310h1fd54f2_1
pyqt5-sip=12.11.0=py310h00ffb61_1
pyqtwebengine=5.15.7=py310h1fd54f2_1
pyrsistent=0.18.1=py310he2412df_1
python=3.10.6=h9a09f29_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pywin32=303=py310he2412df_0
pywinpty=2.0.8=py310h00ffb61_0
pyzmq=24.0.1=py310hcd737a0_0
qt-main=5.15.6=hf0cf448_0
qt-webengine=5.15.4=h7325c83_3
qtconsole=5.3.2=pyhd8ed1ab_0
qtconsole-base=5.3.2=pyha770c72_0
qtpy=2.2.1=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.5.0=pyhd8ed1ab_0
sip=6.7.2=py310h00ffb61_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.39.4=hcfcfb64_0
stack_data=0.5.1=pyhd8ed1ab_0
terminado=0.16.0=pyh08f2357_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
tornado=6.2=py310he2412df_0
traitlets=5.5.0=pyhd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022e=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hac3ee0b_8
vs2015_runtime=14.29.30139=h890b9b1_8
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.4.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=4.0.3=pyhd8ed1ab_0
winpty=0.4.3=4
xz=5.2.6=h8d14728_0
zeromq=4.3.4=h0e60522_1
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h7755175_4
